### 2019-11-17 982defd
actually use roxygen 7 and R6 syntax
