+++ man/LearnerClassif.Rd
@@ -1,152 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/LearnerClassif.R
-\docType{data}
 \name{LearnerClassif}
 \alias{LearnerClassif}
 \title{Classification Learner}
-\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Learner}.}
 \description{
 This Learner specializes \link{Learner} for classification problems.
 Many predefined learners can be found in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link{mlr_learners} after loading the \CRANpkg{mlr3learners} package.
 }
-\section{Construction}{
-\preformatted{l = LearnerClassif$new(id, param_set = ParamSet$new(), predict_types = character(), feature_types = character(),
-    properties = character(), data_formats = "data.table", packages = character(), man = NA_character_)
-}
-
-For a description of the arguments, see \link{Learner}.
-\code{task_type} is set to \code{"classif"}.
-
-Possible values for \code{predict_types} are passed to and converted by \link{PredictionClassif}:
-\itemize{
-\item \code{"response"}: Predicts a class label for each observation in the test set.
-\item \code{"prob"}: Predicts the posterior probability for each class for each observation in the test set.
-}
-
-Additional learner properties include:
-\itemize{
-\item \code{"twoclass"}: The learner works on binary classification problems.
-\item \code{"multiclass"}: The learner works on multiclass classification problems.
-}
-}
-
-\section{Fields}{
-
-\itemize{
-\item \code{id} :: \code{character(1)}\cr
-Identifier of the learner.
-\item \code{task_type} :: \code{character(1)}\cr
-Stores the type of class this learner can operate on, e.g. \code{"classif"} or \code{"regr"}.
-A complete list of task types is stored in \code{\link[=mlr_reflections]{mlr_reflections$task_types}}.
-\item \code{param_set} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr
-Description of available hyperparameters and hyperparameter settings.
-\item \code{predict_types} :: \code{character()}\cr
-Stores the possible predict types the learner is capable of.
-A complete list of candidate predict types, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$learner_predict_types}}.
-\item \code{predict_type} :: \code{character(1)}\cr
-Stores the currently selected predict type. Must be an element of \code{l$predict_types}.
-\item \code{feature_types} :: \code{character()}\cr
-Stores the feature types the learner can handle, e.g. \code{"logical"}, \code{"numeric"}, or \code{"factor"}.
-A complete list of candidate feature types, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$task_feature_types}}.
-\item \code{properties} :: \code{character()}\cr
-Stores a set of properties/capabilities the learner has.
-A complete list of candidate properties, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$learner_properties}}.
-\item \code{packages} :: \code{character()}\cr
-Stores the names of required packages.
-\item \code{state} :: \code{NULL} | named \code{list()}\cr
-Current (internal) state of the learner.
-Contains all information learnt during \code{train()} and \code{predict()}.
-Do not access elements from here directly.
-\item \code{encapsulate} (named \code{character()})\cr
-How to call the code in \code{train_internal()} and \code{predict_internal()}.
-Must be a named character vector with names \code{"train"} and \code{"predict"}.
[…]
