+++ DESCRIPTION
@@ -81,7 +81,7 @@ Encoding: UTF-8
 LazyData: true
 NeedsCompilation: no
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 6.1.1
+RoxygenNote: 7.0.0
 Collate:
     'mlr_reflections.R'
     'BenchmarkResult.R'
