### 2019-11-17 21b4df2
new roxygen2 (for now w/o r6 support)
