+++ man/TaskSupervised.Rd
@@ -1,6 +1,5 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/TaskSupervised.R
-\docType{data}
 \name{TaskSupervised}
 \alias{TaskSupervised}
 \title{Supervised Task}
@@ -45,9 +44,11 @@ Defaults to all rows with role "use".
 task = TaskSupervised$new("iris", task_type = "classif", backend = iris, target = "Species")
 }
 \seealso{
-Other Task: \code{\link{TaskClassif}},
-  \code{\link{TaskRegr}}, \code{\link{Task}},
-  \code{\link{mlr_tasks}}
+Other Task: 
+\code{\link{TaskClassif}},
+\code{\link{TaskRegr}},
+\code{\link{Task}},
+\code{\link{mlr_tasks}}
 }
 \concept{Task}
 \keyword{internal}
