+++ man/LearnerClassifFeatureless.Rd
@@ -1,6 +1,5 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/LearnerClassifFeatureless.R
-\docType{data}
 \name{LearnerClassifFeatureless}
 \alias{LearnerClassifFeatureless}
 \alias{mlr_learners_classif.featureless}
@@ -27,4 +26,3 @@ lrn("classif.featureless")
 
 \code{as.data.table(mlr_learners)} for a complete table of all (also dynamically created) \link{Learner} implementations.
 }
-\keyword{datasets}
