+++ man/MeasureDebug.Rd
@@ -1,6 +1,5 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/MeasureDebug.R
-\docType{data}
 \name{MeasureDebug}
 \alias{MeasureDebug}
 \alias{mlr_measures_debug}
@@ -41,4 +40,3 @@ rr$score(measure)
 
 \code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.
 }
-\keyword{datasets}
