+++ man/DataBackendDataTable.Rd
@@ -1,6 +1,5 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/DataBackendDataTable.R
-\docType{data}
 \name{DataBackendDataTable}
 \alias{DataBackendDataTable}
 \title{DataBackend for data.table}
@@ -51,8 +50,9 @@ b$colnames
 as_data_backend(iris)
 }
 \seealso{
-Other DataBackend: \code{\link{DataBackendMatrix}},
-  \code{\link{DataBackend}}, \code{\link{as_data_backend}}
+Other DataBackend: 
+\code{\link{DataBackendMatrix}},
+\code{\link{DataBackend}},
+\code{\link{as_data_backend}()}
 }
 \concept{DataBackend}
-\keyword{datasets}
