+++ man/Task.Rd
@@ -1,6 +1,5 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/Task.R
-\docType{data}
 \name{Task}
 \alias{Task}
 \title{Task Class}
@@ -51,7 +50,7 @@ Each column (feature) can have an arbitrary number of the following roles:
 \item \code{"feature"}: Regular feature used in the model fitting process.
 \item \code{"target"}: Target variable.
 \item \code{"name"}: Row names / observation labels. To be used in plots.
-\item \code{"order"}: Data returned by \code{$data()} is ordered by this column (or these columns).
+\item \code{"order"}: Data returned by \verb{$data()} is ordered by this column (or these columns).
 \item \code{"group"}: During resampling, observations with the same value of the variable with role "group"
 are marked as "belonging together". They will be exclusively assigned to be either in the training set
 or in the test set for each resampling iteration. Only up to one column may have this role.
@@ -98,7 +97,7 @@ The following properties are currently standardized and understood by tasks in \
 \item \code{"strata"}: The task is resampled using one or more stratification variables (role \code{"stratum"}).
 \item \code{"groups"}: The task comes with grouping/blocking information (role \code{"group"}).
 \item \code{"weights"}: The task comes with observation weights (role \code{"weight"}).
-Note that above listed properties are calculated from the \code{$col_roles} and must not be set explicitly.
+Note that above listed properties are calculated from the \verb{$col_roles} and must not be set explicitly.
 }
 \item \code{strata} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
 If the task has columns designated with role \code{"stratum"}, returns a table with one subpopulation per row and two columns:
@@ -118,7 +117,7 @@ If the task has a column with designated role "weight", table with two columns:
 \code{row_id} (\code{integer()} | \code{character()}) and the observation weights \code{weight} (\code{numeric()}).
 Returns \code{NULL} if there are is no weight column.
 \item \code{man} :: \code{character(1)}\cr
-String in the format \code{[pkg]::[topic]} pointing to a manual page for this object.
+String in the format \verb{[pkg]::[topic]} pointing to a manual page for this object.
 }
 }
 
@@ -135,7 +134,7 @@ columns are filtered to only contain features with roles "target" and "feature".
 If invalid \code{rows} or \code{cols} are specified, an exception is raised.
 \item \code{formula(rhs = ".")}\cr
 \code{character()} -> \code{\link[stats:formula]{stats::formula()}}\cr
-Constructs a \code{\link[stats:formula]{stats::formula()}}, e.g. \code{[target] ~ [feature_1] + [feature_2] + ... + [feature_k]}, using
+Constructs a \code{\link[stats:formula]{stats::formula()}}, e.g. \verb{[target] ~ [feature_1] + [feature_2] + ... + [feature_k]}, using
 the features provided in argument \code{rhs} (defaults to all columns with role \code{"feature"}, symbolized by \code{"."}).
 \item \code{levels(cols = NULL)}\cr
 \code{character()} -> named \code{list()}\cr
@@ -143,7 +142,7 @@ Returns the distinct values for columns referenced in \code{cols} with storage t
 Argument \code{cols} defaults to all such columns with role \code{"target"} or \code{"feature"}.
 
 Note that this function ignores the row roles, it returns all levels available in the \link{DataBackend}.
-To update the stored level information, e.g. after filtering a task, call \code{$droplevels()}.
+To update the stored level information, e.g. after filtering a task, call \verb{$droplevels()}.
 \item \code{droplevels(cols = NULL)}\cr
[…]
