+++ man/Measure.Rd
@@ -1,6 +1,5 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/Measure.R
-\docType{data}
 \name{Measure}
 \alias{Measure}
 \title{Measure Class}
@@ -47,7 +46,7 @@ function defined in \code{aggregator} to average them to a single number.
 
 If set to \code{"micro"}, the individual \link{Prediction} objects are first combined into a single new \link{Prediction} object which is then used to assess the performance.
 The function \code{aggregator} is not used in this case.
-\item \code{aggregator} :: \code{function(x)}\cr
+\item \code{aggregator} :: \verb{function(x)}\cr
 Function to aggregate individual performance scores \code{x} where \code{x} is a numeric vector.
 If \code{NULL}, defaults to \code{\link[=mean]{mean()}}.
 \item \code{properties} :: \code{character()}\cr
@@ -75,7 +74,7 @@ Required task properties, see \link{Task}.
 Set of required packages.
 Note that these packages will be loaded via \code{\link[=requireNamespace]{requireNamespace()}}, and are not attached.
 \item \code{man} :: \code{character(1)}\cr
-String in the format \code{[pkg]::[topic]} pointing to a manual page for this object.
+String in the format \verb{[pkg]::[topic]} pointing to a manual page for this object.
 }
 }
 
@@ -100,13 +99,14 @@ pass the respective \link{Task}, the trained \link{Learner} or the training set 
 This is handled internally during \code{\link[=resample]{resample()}}/\code{\link[=benchmark]{benchmark()}}.
 \item \code{help()}\cr
 () -> \code{NULL}\cr
-Opens the corresponding help page referenced by \code{$man}.
+Opens the corresponding help page referenced by \verb{$man}.
 }
 }
 
 \seealso{
-Other Measure: \code{\link{MeasureClassif}},
-  \code{\link{MeasureRegr}}, \code{\link{mlr_measures}}
+Other Measure: 
+\code{\link{MeasureClassif}},
+\code{\link{MeasureRegr}},
+\code{\link{mlr_measures}}
 }
 \concept{Measure}
-\keyword{datasets}
