+++ man/LearnerRegrFeatureless.Rd
@@ -1,6 +1,5 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/LearnerRegrFeatureless.R
-\docType{data}
 \name{LearnerRegrFeatureless}
 \alias{LearnerRegrFeatureless}
 \alias{mlr_learners_regr.featureless}
@@ -25,4 +24,3 @@ lrn("regr.featureless")
 
 \code{as.data.table(mlr_learners)} for a complete table of all (also dynamically created) \link{Learner} implementations.
 }
-\keyword{datasets}

[…]
