### 2019-11-18 eed0e64
pkgdown: add search
