### 2019-11-18 9f95dea
add condense prototype
