+++ R/resample.R
@@ -6,8 +6,12 @@
 #' @param task :: [Task].
 #' @param learner :: [Learner].
 #' @param resampling :: [Resampling].
-#' @param store_models :: `logical(1)`\cr
+#' @param store_models :: `character(1)` | `logical(1)`\cr
 #'   Keep the fitted model after the test set has been predicted?
+#'   Possible values are `"yes"`, `"no"` (remove the model after prediction) and
+#'   `"condense"` (cut off unimportant parts of the model, see [Learner]).
+#'   For backward compatibility, it is also possible to provide a logical flag here
+#'   (`TRUE` for `"yes"`, `FALSE` for `"no"`).
 #'   Set to `TRUE` if you want to further analyse the models or want to
 #'   extract information like variable importance.
 #' @return [ResampleResult].
@@ -15,10 +19,7 @@
 #'
 #' @template section_parallelization
 #' @template section_logging
-#'
-#' @note
-#' The fitted models are discarded after the predictions have been scored in order to reduce memory consumption.
-#' If you need access to the models for later analysis, set `store_models` to `TRUE`.
+#' @template note_store_models
 #'
 #' @export
 #' @examples
@@ -48,11 +49,17 @@
 #' bmr1 = as_benchmark_result(rr)
 #' bmr2 = as_benchmark_result(rr_featureless)
 #' print(bmr1$combine(bmr2))
-resample = function(task, learner, resampling, store_models = FALSE) {
+resample = function(task, learner, resampling, store_models = "no") {
   task = assert_task(as_task(task, clone = TRUE))
   learner = assert_learner(as_learner(learner, clone = TRUE))
   resampling = assert_resampling(as_resampling(resampling))
-  assert_flag(store_models)
+  if (is.logical(store_models)) {
+    # TODO: deprecate this in the future
+    assert_flag(store_models)
+    store_models = c("no", "yes")[store_models + 1L]
+  } else {
+    assert_choice(store_models, c("no", "yes", "condense"))
+  }
   assert_learnable(task, learner)
 
   instance = resampling$clone(deep = TRUE)
