+++ man-roxygen/note_store_models.R
@@ -1,0 +1,3 @@
+#' @note
+#' The fitted models are discarded after the predictions have been scored in order to reduce memory consumption.
+#' If you need access to the models for later analysis, set `store_models` to `"yes"` or `"condense".
