### 2019-11-19 #1893 @terrytangyuan
@mllg Thanks for the update. If you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
