### 2019-11-19 #1893 @mllg
After skimming through the list filtering for R as primary programming language and machine learning as topic, these seem to be potential reviewers (in no particular order):

@robohyun66
@osorensen
@expectopatronum
@nhejazi
