### 2019-11-19 #1893 @nhejazi
Looks quite interesting! Happy to provide a review if helpful.
