### 2019-11-19 #1893 @mllg
I took the opportunity to make one last minor change to the article. I'm wondering if I can also trigger the bot to re-create the article proof?
