+++ paper/paper.md
@@ -1,0 +1,102 @@
+---
+title: 'mlr3: A modern object-oriented machine learning framework in R'
+tags:
+  - R
+  - machine learning
+  - classification
+  - regression
+authors:
+  - name: Michel Lang
+    orcid: 0000-0001-9754-0393
+    affiliation: "1, 2"
+  - name: Martin Binder
+    affiliation: 2
+  - name: Jakob Richter
+    orcid: 0000-0003-4481-5554
+    affiliation: 1
+  - name: Patrick Schratz
+    orcid: 0000-0003-0748-6624
+    affiliation: 2
+  - name: Florian Pfisterer
+    orcid: 0000-0001-8867-762X
+    affiliation: 2
+  - name: Stefan Coors
+    orcid: 0000-0002-7465-2146
+    affiliation: 2
+  - name: Quay Au
+    orcid: 0000-0002-5252-8902
+    affiliation: 2
+  - name: Giuseppe Casalicchio
+    orcid: 0000-0001-5324-5966
+    affiliation: 2
+  - name: Lars Kotthoff
+    orcid: 0000-0003-4635-6873
+    affiliation: 3
+  - name: Bernd Bischl
+    orcid: 0000-0001-6002-6980
+    affiliation: 2
+affiliations:
+ - name: TU Dortmund University
+   index: 1
+ - name: LMU Munich
+   index: 2
+ - name: University of Wyoming
+   index: 3
+date: 1 December 2019
+bibliography: paper.bib
+---
+
+# Summary
+
+The [`R`](https://www.r-project.org/) [@R] package [`mlr3`](https://mlr3.mlr-org.com) and its associated ecosystem of extension packages implements a powerful, object-oriented and extensible framework for machine learning (ML) in `R`.
+It provides a unified interface to many learning algorithms available on [CRAN](https://cran.r-project.org/), augmenting them with model-agnostic general-purpose functionality that is needed in every ML project, for example train-test-evaluation, resampling, preprocessing, hyperparameter tuning, nested resampling, and visualization of results from ML experiments.
+The package is a complete reimplementation of the [`mlr`](https://mlr.mlr-org.com) [@mlr] package that leverages many years of experience and learned best practices to provide a state-of-the-art system that is powerful, flexible, extensible, and maintainable.
+We target both **practitioners** who want to quickly apply ML algorithms to their problems and **researchers** who want to implement, benchmark, and compare their new methods in a structured environment.
+[`mlr3`](https://mlr3.mlr-org.com) is suitable for short scripts that test an idea, for complex multi-stage experiments with advanced functionality that use a broad range of ML functionality, as a foundation to implement new ML (meta-)algorithms (for example AutoML systems), and everything in between.
+Functional correctness is ensured through extensive unit and integration tests.
+
+Several other general-purpose ML toolboxes exist for different programing languages.
+The most widely used ones are [scikit-learn](https://scikit-learn.org/) [@sk-learn] for Python , [Weka](https://www.cs.waikato.ac.nz/ml/weka/) [@weka] for Java, and [mlj](https://github.com/alan-turing-institute/MLJ.jl) [@mlj] for Julia.
+The most important toolboxes for `R` are [mlr](https://cran.r-project.org/package=mlr), [caret](https://cran.r-project.org/package=caret) [@caret] and [tidymodels](https://cran.r-project.org/package=tidymodels) [@tidymodels].
+
+# Lessons Learned from 6 Years of Machine Learning in `R`
+
+The predecessor package [`mlr`](https://mlr.mlr-org.com) was first released to [CRAN](https://cran.r-project.org/) in 2013, with the core design and architecture dating back much further.
+As with most software, more code was added over time to integrate more ML algorithms, more approaches for feature selection or hyperparameter tuning, more methods to analyze trained models, and many other things.
[…]
