### 2019-11-19 #1903 @whedon
[REVIEW]: mlr3: A modern object-oriented machine learning framework in R
**Submitting author:** @mllg (<a href="http://orcid.org/0000-0001-9754-0393">Michel Lang</a>)
**Repository:** <a href="https://github.com/mlr-org/mlr3" target ="_blank">https://github.com/mlr-org/mlr3</a>
**Version:** v0.1.5

@nhejazi , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
