+++ R/DataBackend.R
@@ -115,7 +115,7 @@ DataBackend = R6Class("DataBackend", cloneable = FALSE,
     print = function() {
       nr = self$nrow
       catf("%s (%ix%i)", format(self), nr, self$ncol)
-      print(self$head(6L), row.names = FALSE)
+      print(self$head(6L), row.names = FALSE, print.keys = FALSE)
       if (nr > 6L) {
         catf("[...] (%i rows omitted)", nr - 6L)
       }
