+++ pkgdown.yml
@@ -1,0 +1,8 @@
+pandoc: '2.2'
+pkgdown: 1.4.1
+pkgdown_sha: ~
+articles: []
+urls:
+  reference: https://mlr3.mlr-org.com/reference
+  article: https://mlr3.mlr-org.com/articles
+
