### 2019-11-23 fa52e3f
pkgdown: link mlr3book in navbar
