+++ pkgdown/_pkgdown.yml
@@ -19,13 +19,21 @@ toc:
 
 navbar:
   structure:
-    left:  [reference, news]
+    left:  [reference, news, book]
     right: [github, slack, stackoverflow, rss]
   components:
     home: ~
+    reference:
+      icon: fa fa-file-alt
+      text: Reference
+      href: reference/index.html
     slack:
       icon: fab fa-slack
       href: https://mlr-org.slack.com
+    book:
+      text: mlr3book
+      icon: fa fa-link
+      href: https://mlr3book.mlr-org.com
     stackoverflow:
       icon: fab fa-stack-overflow
       href: https://stackoverflow.com/questions/tagged/mlr
