+++ tic.R
@@ -3,7 +3,7 @@ do_package_checks(error_on = "warning")
 if (ci_has_env("BUILD_PKGDOWN")) {
   get_stage("install") %>%
     add_step(step_install_github("mlr-org/mlr3pkgdowntemplate"))
-  do_pkgdown(orphan = TRUE)
+  do_pkgdown()
 }
 
 get_stage("after_success") %>%
