+++ README.Rmd
@@ -4,6 +4,8 @@ output: github_document
 
 # mlr3 <img src="man/figures/logo.png" align="right" width = "120" />
 
+Package website: [release](https://mlr3.mlr-org.com/) | [dev](https://mlr3.mlr-org.com/dev)
+
 Efficient, object-oriented programming on the building blocks of machine learning.
 Successor of [mlr](https://github.com/mlr-org/mlr).
 
