### 2019-11-24 1b0c365
Deploy from Travis build 1994 [ci skip]
Build URL: https://travis-ci.org/mlr-org/mlr3/builds/616022815
Commit: 56e7a674fe8a1076959a4c24c2c2203f066de389
