+++ dev/pkgdown.yml
@@ -1,6 +1,6 @@
-pandoc: 2.7.3
-pkgdown: 1.4.1.9000
-pkgdown_sha: d0b233e7966dc86c2b315f880deead3c7f9a5a02
+pandoc: '2.2'
+pkgdown: 1.4.1
+pkgdown_sha: ~
 articles: []
 urls:
   reference: https://mlr3.mlr-org.com/reference
