### 2019-11-26 1e1ffc2
improve error message in autotest
