### 2019-11-28 992125a
refactored documentation of measures
