+++ R/MeasureSimple.R
@@ -13,7 +13,7 @@ MeasureBinarySimple = R6Class("MeasureBinaryimple",
         predict_type = info$predict_type,
         task_properties = "twoclass",
         packages = "mlr3measures",
-        man = paste0("mlr3measures::", name)
+        man = paste0("mlr3::mlr_measures_classif.", name)
       )
       self$fun = get(name, envir = asNamespace("mlr3measures"), mode = "function")
     },
@@ -45,7 +45,7 @@ MeasureClassifSimple = R6Class("MeasureClassifSimple",
         minimize = info$minimize,
         predict_type = info$predict_type,
         packages = "mlr3measures",
-        man = paste0("mlr3measures::", name)
+        man = paste0("mlr3::mlr_measures_classif.", name)
       )
       self$fun = get(name, envir = asNamespace("mlr3measures"), mode = "function")
     },
@@ -75,7 +75,7 @@ MeasureRegrSimple = R6Class("MeasureRegrSimple",
         minimize = info$minimize,
         predict_type = info$predict_type,
         packages = "mlr3measures",
-        man = paste0("mlr3measures::", name)
+        man = paste0("mlr3::mlr_measures_regr.", name)
       )
       self$fun = get(name, envir = asNamespace("mlr3measures"), mode = "function")
     },
@@ -91,161 +91,185 @@ MeasureRegrSimple = R6Class("MeasureRegrSimple",
   )
 )
 
+### binary classification measures
 
-#' @title Measures implemented in \CRANpkg{mlr3measures}
-#'
-#' @name MeasureSimple
-#' @usage NULL
-#' @include mlr_measures.R
-#' @aliases
-#'   mlr_measures_classif.tpr
-#'   mlr_measures_classif.auc
-#'   mlr_measures_regr.mse
-#'   mlr_measures_classif.ce
-#'   mlr_measures_regr.rse
-#'   mlr_measures_classif.tn
-#'   mlr_measures_classif.precision
-#'   mlr_measures_classif.tp
-#'   mlr_measures_classif.specificity
-#'   mlr_measures_classif.sensitivity
-#'   mlr_measures_classif.fbeta
-#'   mlr_measures_classif.tnr
-#'   mlr_measures_classif.dor
-#'   mlr_measures_classif.recall
-#'   mlr_measures_regr.msle
-#'   mlr_measures_regr.medae
-#'   mlr_measures_regr.sae
-#'   mlr_measures_regr.rsq
-#'   mlr_measures_classif.fpr
-#'   mlr_measures_regr.sse
-#'   mlr_measures_regr.medse
-#'   mlr_measures_classif.ppv
-#'   mlr_measures_classif.fn
-#'   mlr_measures_classif.fp
-#'   mlr_measures_classif.fdr
-#'   mlr_measures_classif.fomr
-#'   mlr_measures_classif.fnr
-#'   mlr_measures_classif.mcc
-#'   mlr_measures_classif.logloss
-#'   mlr_measures_regr.pbias
-#'   mlr_measures_regr.maxae
-#'   mlr_measures_regr.rmse
-#'   mlr_measures_regr.mape
-#'   mlr_measures_regr.smape
-#'   mlr_measures_regr.rrse
-#'   mlr_measures_classif.npv
-#'   mlr_measures_classif.acc
-#'   mlr_measures_classif.bacc
-#'   mlr_measures_regr.bias
-#'   mlr_measures_regr.maxse
-#'   mlr_measures_regr.rae
-#'   mlr_measures_regr.ktau
-#'   mlr_measures_regr.mae
-#'   mlr_measures_regr.rmsle
-#'   mlr_measures_regr.srho
-#' @format [R6::R6Class()] inheriting from [Measure].
-#'
-#' @section Construction:
[…]
