+++ man/MeasureSimple.Rd
@@ -1,147 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/MeasureSimple.R
-\name{MeasureSimple}
-\alias{MeasureSimple}
-\alias{mlr_measures_classif.tpr}
-\alias{mlr_measures_classif.auc}
-\alias{mlr_measures_regr.mse}
-\alias{mlr_measures_classif.ce}
-\alias{mlr_measures_regr.rse}
-\alias{mlr_measures_classif.tn}
-\alias{mlr_measures_classif.precision}
-\alias{mlr_measures_classif.tp}
-\alias{mlr_measures_classif.specificity}
-\alias{mlr_measures_classif.sensitivity}
-\alias{mlr_measures_classif.fbeta}
-\alias{mlr_measures_classif.tnr}
-\alias{mlr_measures_classif.dor}
-\alias{mlr_measures_classif.recall}
-\alias{mlr_measures_regr.msle}
-\alias{mlr_measures_regr.medae}
-\alias{mlr_measures_regr.sae}
-\alias{mlr_measures_regr.rsq}
-\alias{mlr_measures_classif.fpr}
-\alias{mlr_measures_regr.sse}
-\alias{mlr_measures_regr.medse}
-\alias{mlr_measures_classif.ppv}
-\alias{mlr_measures_classif.fn}
-\alias{mlr_measures_classif.fp}
-\alias{mlr_measures_classif.fdr}
-\alias{mlr_measures_classif.fomr}
-\alias{mlr_measures_classif.fnr}
-\alias{mlr_measures_classif.mcc}
-\alias{mlr_measures_classif.logloss}
-\alias{mlr_measures_regr.pbias}
-\alias{mlr_measures_regr.maxae}
-\alias{mlr_measures_regr.rmse}
-\alias{mlr_measures_regr.mape}
-\alias{mlr_measures_regr.smape}
-\alias{mlr_measures_regr.rrse}
-\alias{mlr_measures_classif.npv}
-\alias{mlr_measures_classif.acc}
-\alias{mlr_measures_classif.bacc}
-\alias{mlr_measures_regr.bias}
-\alias{mlr_measures_regr.maxse}
-\alias{mlr_measures_regr.rae}
-\alias{mlr_measures_regr.ktau}
-\alias{mlr_measures_regr.mae}
-\alias{mlr_measures_regr.rmsle}
-\alias{mlr_measures_regr.srho}
-\title{Measures implemented in \CRANpkg{mlr3measures}}
-\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
-\description{
-The following measures from \CRANpkg{mlr3measures} are interfaced using a simple utility class:
-
-Binary classification (inheriting from \link{MeasureClassif}):
-\itemize{
-\item \link[mlr3measures:auc]{auc}: Area Under the ROC Curve
-\item \link[mlr3measures:dor]{dor}: Diagnostic Odds Ratio
-\item \link[mlr3measures:fbeta]{fbeta}: F-beta score
-\item \link[mlr3measures:fdr]{fdr}: False Discovery Rate
-\item \link[mlr3measures:fn]{fn}: False Negatives
-\item \link[mlr3measures:fnr]{fnr}: False Negative Rate
-\item \link[mlr3measures:fomr]{fomr}: False Omission Rate
-\item \link[mlr3measures:fp]{fp}: False Positives
-\item \link[mlr3measures:fpr]{fpr}: False Positive Rate
-\item \link[mlr3measures:mcc]{mcc}: Matthews Correlation Coefficient
-\item \link[mlr3measures:npv]{npv}: Negative Predictive Value
[…]
