+++ man/mlr_measures_classif.auc.Rd
@@ -1,0 +1,89 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/MeasureSimple.R
+\name{mlr_measures_classif.auc}
+\alias{mlr_measures_classif.auc}
+\title{Area Under the ROC Curve}
+\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
+\description{
+Computes the area under the Receiver Operator Characteristic (ROC) curve.
+The AUC can be interpreted as the probability that a randomly chosen positive observation
+has a higher predicted probability than a randomly chosen negative observation.
+}
+\note{
+The score function calls \code{\link[mlr3measures:auc]{mlr3measures::auc()}} from package \CRANpkg{mlr3measures}.
+
+If the measure is undefined for the input, \code{NaN} is returned.
+This can be customized by setting the field \code{na_value}.
+}
+\section{Construction}{
+
+This measures can be retrieved from the dictionary \link{mlr_measures}:\preformatted{mlr_measures$get("classif.auc")
+msr("classif.auc")
+}
+}
+
+\section{Meta Information}{
+
+
+\itemize{
+\item Type: \code{"binary"}
+\item Range: \eqn{[0, 1]}{[0, 1]}
+\item Minimize: \code{FALSE}
+\item Required prediction: \code{prob}
+}
+
+}
+
+\seealso{
+\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
+
+\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.
+
+Other classification measures: 
+\code{\link{mlr_measures_classif.acc}},
+\code{\link{mlr_measures_classif.bacc}},
+\code{\link{mlr_measures_classif.ce}},
+\code{\link{mlr_measures_classif.dor}},
+\code{\link{mlr_measures_classif.fbeta}},
+\code{\link{mlr_measures_classif.fdr}},
+\code{\link{mlr_measures_classif.fnr}},
+\code{\link{mlr_measures_classif.fn}},
+\code{\link{mlr_measures_classif.fomr}},
+\code{\link{mlr_measures_classif.fpr}},
+\code{\link{mlr_measures_classif.fp}},
+\code{\link{mlr_measures_classif.logloss}},
+\code{\link{mlr_measures_classif.mcc}},
+\code{\link{mlr_measures_classif.npv}},
+\code{\link{mlr_measures_classif.ppv}},
+\code{\link{mlr_measures_classif.precision}},
+\code{\link{mlr_measures_classif.recall}},
+\code{\link{mlr_measures_classif.sensitivity}},
+\code{\link{mlr_measures_classif.specificity}},
+\code{\link{mlr_measures_classif.tnr}},
+\code{\link{mlr_measures_classif.tn}},
+\code{\link{mlr_measures_classif.tpr}},
+\code{\link{mlr_measures_classif.tp}}
+
+Other binary classification measures: 
+\code{\link{mlr_measures_classif.dor}},
+\code{\link{mlr_measures_classif.fbeta}},
+\code{\link{mlr_measures_classif.fdr}},
+\code{\link{mlr_measures_classif.fnr}},
+\code{\link{mlr_measures_classif.fn}},
+\code{\link{mlr_measures_classif.fomr}},
+\code{\link{mlr_measures_classif.fpr}},
+\code{\link{mlr_measures_classif.fp}},
+\code{\link{mlr_measures_classif.mcc}},
+\code{\link{mlr_measures_classif.npv}},
+\code{\link{mlr_measures_classif.ppv}},
[…]
