### 2019-11-28 bfa86c4
rename man pages of measures
