+++ R/MeasureSelectedFeatures.R
@@ -1,7 +1,7 @@
 #' @title Selected Features Measure
 #'
 #' @usage NULL
-#' @aliases mlr_measures_selected_features
+#' @name mlr_measures_selected_features
 #' @format [R6::R6Class()] inheriting from [Measure].
 #' @include Measure.R
 #'
