+++ man/mlr_measures_selected_features.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/MeasureSelectedFeatures.R
-\name{MeasureSelectedFeatures}
-\alias{MeasureSelectedFeatures}
+\name{mlr_measures_selected_features}
 \alias{mlr_measures_selected_features}
+\alias{MeasureSelectedFeatures}
 \title{Selected Features Measure}
 \format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
 \description{
