+++ man/mlr_measures_oob_error.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/MeasureOOBError.R
-\name{MeasureOOBError}
-\alias{MeasureOOBError}
+\name{mlr_measures_oob_error}
 \alias{mlr_measures_oob_error}
+\alias{MeasureOOBError}
 \title{Out-of-bag Error Measure}
 \format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
 \description{
