+++ man/mlr_measures_debug.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/MeasureDebug.R
-\name{MeasureDebug}
-\alias{MeasureDebug}
+\name{mlr_measures_debug}
 \alias{mlr_measures_debug}
+\alias{MeasureDebug}
 \title{Debug Measure}
 \format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
 \description{
