+++ man/mlr_measures_classif.costs.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/MeasureClassifCosts.R
-\name{MeasureClassifCosts}
-\alias{MeasureClassifCosts}
+\name{mlr_measures_classif.costs}
 \alias{mlr_measures_classif.costs}
+\alias{MeasureClassifCosts}
 \title{Cost-sensitive Classification Measure}
 \format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
 \description{
