+++ man/mlr_measures_elapsed_time.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/MeasureElapsedTime.R
-\name{MeasureElapsedTime}
-\alias{MeasureElapsedTime}
+\name{mlr_measures_elapsed_time}
 \alias{mlr_measures_elapsed_time}
+\alias{MeasureElapsedTime}
 \alias{mlr_measures_time_train}
 \alias{mlr_measures_time_predict}
 \alias{mlr_measures_time_both}
