### 2019-11-28 2575949
improve measure docs
