+++ R/MeasureOOBError.R
@@ -5,6 +5,12 @@
 #' @format [R6::R6Class()] inheriting from [Measure].
 #' @include Measure.R
 #'
+#' @description
+#' Returns the out-of-bag error of the [Learner] for learners that support it
+#' (learners with property `"oob_error"`).
+#' Returns `NA` for unsupported learners.
+#'
+#'
 #' @section Construction:
 #' ```
 #' MeasureOOBError$new()
@@ -12,10 +18,11 @@
 #' msr("oob_error")
 #' ```
 #'
-#' @description
-#' Returns the out-of-bag error of the [Learner] for learners that support it
-#' (learners with property `"oob_error"`).
-#' Returns `NA` for unsupported learners.
+#' @section Meta Information:
+#' * Type: `NA`
+#' * Range: \eqn{(-\infty, \infty)}{(-Inf, Inf)}
+#' * Minimize: `NA`
+#' * Required prediction: 'response'
 #'
 #' @template seealso_measure
 #' @export
