+++ R/MeasureDebug.R
@@ -5,6 +5,10 @@
 #' @format [R6::R6Class()] inheriting from [Measure].
 #' @include Measure.R
 #'
+#' @description
+#' This measure returns the number of observations in the [Prediction] object.
+#' Its main purpose is debugging.
+#'
 #' @section Construction:
 #' ```
 #' MeasureDebug$new(na_ratio = 0)
@@ -16,12 +20,14 @@
 #'   Ratio of scores which should be `NA`.
 #'   Default is 0.
 #'
-#' @description
-#' This measure returns the number of observations in the [Prediction] object.
-#' Its main purpose is debugging.
-#'
 #' @section Fields:
 #' * `na_ratio` :: `numeric(1)`.
+#'
+#' @section Meta Information:
+#' * Type: `NA`
+#' * Range: \eqn{[0, \infty)}{[0, Inf)}
+#' * Minimize: `NA`
+#' * Required prediction: 'response'
 #'
 #' @template seealso_measure
 #' @export
