+++ man/mlr_measures_classif.ce.Rd
@@ -45,6 +45,7 @@ Other classification measures:
 \code{\link{mlr_measures_classif.acc}},
 \code{\link{mlr_measures_classif.auc}},
 \code{\link{mlr_measures_classif.bacc}},
+\code{\link{mlr_measures_classif.costs}},
 \code{\link{mlr_measures_classif.dor}},
 \code{\link{mlr_measures_classif.fbeta}},
 \code{\link{mlr_measures_classif.fdr}},
@@ -69,6 +70,7 @@ Other classification measures:
 Other multiclass classification measures: 
 \code{\link{mlr_measures_classif.acc}},
 \code{\link{mlr_measures_classif.bacc}},
+\code{\link{mlr_measures_classif.costs}},
 \code{\link{mlr_measures_classif.logloss}}
 }
 \concept{classification measures}
