### 2019-12-01 #1903 @osorensen
@mllg, the paper is very well written. My only suggestion is to correct the spelling of the word _techniques_ in the sentence:

> ... integrates many feature filtering technqiues and ...
