### 2019-12-01 #1903 @mllg
Thanks, fixed in https://github.com/mlr-org/mlr3/commit/bd32569c234a88a2bb740e26975adaa487eb0de2.
