+++ R/Learner.R
@@ -1,14 +1,9 @@
-#' @title Learner Class
-#'
-#' @usage NULL
-#' @format [R6::R6Class] object.
-#' @include mlr_reflections.R
+#' Learner
 #'
 #' @description
 #' This is the abstract base class for learner objects like [LearnerClassif] and [LearnerRegr].
 #'
 #' Learners consist of the following parts:
-#'
 #' - Methods `train()` and `predict()` which call `train_internal()` and `predict_internal()`.
 #' - The fitted model, after calling `train()`.
 #' - A [paradox::ParamSet] which stores meta-information about available hyperparameters, and also stores hyperparameter settings.
@@ -17,145 +12,6 @@
 #' Predefined learners are stored in the [mlr3misc::Dictionary] [mlr_learners],
 #' e.g. [`classif.rpart`][mlr_learners_classif.rpart] or [`regr.rpart`][mlr_learners_regr.rpart].
 #' A guide on how to extend \CRANpkg{mlr3} with custom learners can be found in the [mlr3book](https://mlr3book.mlr-org.com).
-#'
-#'
-#' @section Construction:
-#' Note: This object is typically constructed via a derived classes, e.g. [LearnerClassif] or [LearnerRegr].
-#'
-#' ```
-#' l = Learner$new(id, task_type, param_set = ParamSet$new(), predict_types = character(),
-#'      feature_types = character(), properties = character(), packages = character())
-#' ```
-#'
-#' * `id` :: `character(1)`\cr
-#'   Identifier for the learner.
-#'
-#' * `task_type` :: `character(1)`\cr
-#'   Type of the task the learner can operator on. E.g., `"classif"` or `"regr"`.
-#'
-#' * `param_set` :: [paradox::ParamSet]\cr
-#'   Set of hyperparameters.
-#'
-#' * `predict_types` :: `character()`\cr
-#'   Supported predict types. Must be a subset of [`mlr_reflections$learner_predict_types`][mlr_reflections].
-#'
-#' * `predict_sets` :: `character()`\cr
-#'   Sets to predict on during [resample()]/[benchmark()].
-#'   Creates and stores a separate [Prediction] object for each set.
-#'   The individual sets can be combined via getters in [ResampleResult]/[BenchmarkResult], or [Measure]s can be set to operate on subsets of the calculated prediction sets.
-#'   Must be a non-empty subset of `("train", "test")`.
-#'   Default is `"test"`.
-#'
-#' * `feature_types` :: `character()`\cr
-#'   Feature types the learner operates on. Must be a subset of [`mlr_reflections$task_feature_types`][mlr_reflections].
-#'
-#' * `properties` :: `character()`\cr
-#'   Set of properties of the learner. Must be a subset of [`mlr_reflections$learner_properties`][mlr_reflections].
-#'   The following properties are currently standardized and understood by learners in \CRANpkg{mlr3}:
-#'   * `"missings"`: The learner can handle missing values in the data.
-#'   * `"weights"`: The learner supports observation weights.
-#'   * `"importance"`: The learner supports extraction of importance scores, i.e. comes with a `importance()` extractor function (see section on optional extractors).
-#'   * `"selected_features"`: The learner supports extraction of the set of selected features, i.e. comes with a `selected_features()` extractor function (see section on optional extractors).
-#'   * `"oob_error"`: The learner supports extraction of estimated out of bag error, i.e. comes with a `oob_error()` extractor function (see section on optional extractors).
-#'
-#' * `data_formats` :: `character()`\cr
-#'   Vector of supported data formats which can be processed during `$train()` and `$predict()`.
-#'   Defaults to `"data.table"`.
-#'
-#' * `packages` :: `character()`\cr
-#'   Set of required packages.
[…]
