+++ man/Learner.Rd
@@ -2,8 +2,7 @@
 % Please edit documentation in R/Learner.R
 \name{Learner}
 \alias{Learner}
-\title{Learner Class}
-\format{\link[R6:R6Class]{R6::R6Class} object.}
+\title{Learner}
 \description{
 This is the abstract base class for learner objects like \link{LearnerClassif} and \link{LearnerRegr}.
 
@@ -19,126 +18,6 @@ Predefined learners are stored in the \link[mlr3misc:Dictionary]{mlr3misc::Dicti
 e.g. \code{\link[=mlr_learners_classif.rpart]{classif.rpart}} or \code{\link[=mlr_learners_regr.rpart]{regr.rpart}}.
 A guide on how to extend \CRANpkg{mlr3} with custom learners can be found in the \href{https://mlr3book.mlr-org.com}{mlr3book}.
 }
-\section{Construction}{
-
-Note: This object is typically constructed via a derived classes, e.g. \link{LearnerClassif} or \link{LearnerRegr}.\preformatted{l = Learner$new(id, task_type, param_set = ParamSet$new(), predict_types = character(),
-     feature_types = character(), properties = character(), packages = character())
-}
-\itemize{
-\item \code{id} :: \code{character(1)}\cr
-Identifier for the learner.
-\item \code{task_type} :: \code{character(1)}\cr
-Type of the task the learner can operator on. E.g., \code{"classif"} or \code{"regr"}.
-\item \code{param_set} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr
-Set of hyperparameters.
-\item \code{predict_types} :: \code{character()}\cr
-Supported predict types. Must be a subset of \code{\link[=mlr_reflections]{mlr_reflections$learner_predict_types}}.
-\item \code{predict_sets} :: \code{character()}\cr
-Sets to predict on during \code{\link[=resample]{resample()}}/\code{\link[=benchmark]{benchmark()}}.
-Creates and stores a separate \link{Prediction} object for each set.
-The individual sets can be combined via getters in \link{ResampleResult}/\link{BenchmarkResult}, or \link{Measure}s can be set to operate on subsets of the calculated prediction sets.
-Must be a non-empty subset of \verb{("train", "test")}.
-Default is \code{"test"}.
-\item \code{feature_types} :: \code{character()}\cr
-Feature types the learner operates on. Must be a subset of \code{\link[=mlr_reflections]{mlr_reflections$task_feature_types}}.
-\item \code{properties} :: \code{character()}\cr
-Set of properties of the learner. Must be a subset of \code{\link[=mlr_reflections]{mlr_reflections$learner_properties}}.
-The following properties are currently standardized and understood by learners in \CRANpkg{mlr3}:
-\itemize{
-\item \code{"missings"}: The learner can handle missing values in the data.
-\item \code{"weights"}: The learner supports observation weights.
-\item \code{"importance"}: The learner supports extraction of importance scores, i.e. comes with a \code{importance()} extractor function (see section on optional extractors).
-\item \code{"selected_features"}: The learner supports extraction of the set of selected features, i.e. comes with a \code{selected_features()} extractor function (see section on optional extractors).
-\item \code{"oob_error"}: The learner supports extraction of estimated out of bag error, i.e. comes with a \code{oob_error()} extractor function (see section on optional extractors).
-}
-\item \code{data_formats} :: \code{character()}\cr
-Vector of supported data formats which can be processed during \verb{$train()} and \verb{$predict()}.
-Defaults to \code{"data.table"}.
[…]
