+++ DESCRIPTION
@@ -80,7 +80,7 @@ RdMacros: mlr3misc
 Encoding: UTF-8
 LazyData: true
 NeedsCompilation: no
-Roxygen: list(markdown = TRUE, r6 = FALSE)
+Roxygen: list(markdown = TRUE, r6 = TRUE)
 RoxygenNote: 7.0.2
 Collate:
     'mlr_reflections.R'

[…]
