### 2019-12-06 #1903 @nhejazi
@mllg, the `mlr3` R package looks quite useful, well-tested, and well-documented (from a design perspective, our own `sl3` package https://github.com/tlverse/sl3/ seems to have tried to do something very similar, so it's good to see a more robust project on the same track). To echo @osorensen, the paper is also quite excellently written. You've made the job of reviewing this work very easy!
