+++ DESCRIPTION
@@ -81,7 +81,7 @@ Encoding: UTF-8
 LazyData: true
 NeedsCompilation: no
 Roxygen: list(markdown = TRUE, r6 = FALSE)
-RoxygenNote: 7.0.1
+RoxygenNote: 7.0.2
 Collate:
     'mlr_reflections.R'
     'BenchmarkResult.R'
