### 2019-12-10 645ac6c
cleanup NEWS
