### 2019-12-10 #1903 @mllg
I've released version 0.1.5 on GitHub (https://github.com/mlr-org/mlr3/releases/tag/v0.1.5) and CRAN (https://cran.r-project.org/package=mlr3). The GitHub release is uploaded to Zenodo with DOI 10.5281/zenodo.3569963: https://zenodo.org/record/3569963
