### 2019-12-11 0007620
fix assertion in PredictionClassif
